requirements.txt
# ============================================
# CRM API - Dependencias del Proyecto
# ============================================
# Instalar con: pip install -r requirements.txt

# ============================================
# Core Django
# ============================================
Django==4.2.7

# ============================================
# Django REST Framework
# ============================================
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0

# ============================================
# Documentación API (Swagger/OpenAPI)
# ============================================
drf-spectacular==0.26.5

# ============================================
# Validación de Datos
# ============================================
pydantic==2.11.9

# ============================================
# Base de Datos
# ============================================
PyMySQL==1.1.2

# ============================================
# Configuración y Variables de Entorno
# ============================================
django-environ==0.12.0

# ============================================
# CORS (Cross-Origin Resource Sharing)
# ============================================
django-cors-headers==4.7.0

# ============================================
# Filtros y Búsqueda
# ============================================
django-filter==23.3

# ============================================
# Servidor WSGI para Producción
# ============================================
gunicorn==21.2.0

